+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next Release
 
-- Add `retrieveStatelessRates` function and helper function `getLowestStatlessRate`
+- Add `retrieveStatelessRates` function and helper function `getLowestStatelessRate`
 
 ## v6.2.0 (2023-01-18)